+++ src/Http/Requests/JsonApiRequest.php
@@ -24,15 +24,15 @@ use CloudCreativity\LaravelJsonApi\Exceptions\InvalidJsonException;
 use CloudCreativity\LaravelJsonApi\Exceptions\RuntimeException;
 use CloudCreativity\LaravelJsonApi\Object\ResourceIdentifier;
 use CloudCreativity\LaravelJsonApi\Routing\ResourceRegistrar;
+use Illuminate\Contracts\Container\Container;
 use Illuminate\Http\Request;
 use Neomerx\JsonApi\Contracts\Encoder\Parameters\EncodingParametersInterface;
-use Neomerx\JsonApi\Contracts\Http\HttpFactoryInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use function CloudCreativity\LaravelJsonApi\http_contains_body;
 use function CloudCreativity\LaravelJsonApi\json_decode;
 
 /**
- * Class IlluminateRequest
+ * Class JsonApiRequest
  *
  * @package CloudCreativity\LaravelJsonApi
  */
@@ -45,19 +45,14 @@ class JsonApiRequest
     private $request;
 
     /**
-     * @var ServerRequestInterface
-     */
-    private $serverRequest;
-
-    /**
-     * @var HttpFactoryInterface
-     */
-    private $factory;
-
-    /**
      * @var ResolverInterface
      */
     private $resolver;
+
+    /**
+     * @var Container
+     */
+    private $container;
 
     /**
      * @var string|null
@@ -83,20 +78,14 @@ class JsonApiRequest
      * IlluminateRequest constructor.
      *
      * @param Request $request
-     * @param ServerRequestInterface $serverRequest
      * @param ResolverInterface $resolver
-     * @param HttpFactoryInterface $factory
+     * @param Container $container
      */
-    public function __construct(
-        Request $request,
-        ServerRequestInterface $serverRequest,
-        ResolverInterface $resolver,
-        HttpFactoryInterface $factory
-    ) {
+    public function __construct(Request $request, ResolverInterface $resolver, Container $container)
+    {
         $this->request = $request;
-        $this->serverRequest = $serverRequest;
         $this->resolver = $resolver;
-        $this->factory = $factory;
+        $this->container = $container;
     }
 
     /**
@@ -256,7 +245,7 @@ class JsonApiRequest
             return $this->parameters;
         }
 
-        return $this->parameters = $this->parseParameters();
+        return $this->parameters = $this->container->make(EncodingParametersInterface::class);
     }
 
     /**
@@ -526,21 +515,14 @@ class JsonApiRequest
             return false;
         }
 
-        if (!http_contains_body($this->serverRequest)) {
+        $serverRequest = $this->container->make(ServerRequestInterface::class);
+
+        /** @todo allow a Laravel request to be passed to http_contains_body */
+        if (!http_contains_body($serverRequest)) {
             return false;
         }
 
-        return json_decode((string) $this->serverRequest->getBody());
-    }
-
-    /**
-     * @return EncodingParametersInterface
-     */
-    private function parseParameters(): EncodingParametersInterface
-    {
-        $parser = $this->factory->createQueryParametersParser();
-
-        return $parser->parseQueryParameters($this->serverRequest->getQueryParams());
+        return json_decode($this->request->getContent());
     }
 
     /**